+++ config/firebase-config.js
@@ -8,6 +8,12 @@
  */
 window.firebaseConfig = Object.freeze({
     databaseURL: "https://realtime-database-9fc52-default-rtdb.asia-southeast1.firebasedatabase.app",
-    authToken: "",
-    requestTimeoutMs: 15000
+    requestTimeoutMs: 15000,
+    auth: Object.freeze({
+        mode: "firebase-email-password",
+        apiKey: "",
+        identityToolkitURL: "https://identitytoolkit.googleapis.com/v1",
+        secureTokenURL: "https://securetoken.googleapis.com/v1",
+        refreshSkewSeconds: 120
+    })
 });